pyproject.toml
[project]
requires-python = ">= 3.8"
name = "omnimirror"
version = "0.1.0"
dependencies = ["click", "python-debian"]

[project.scripts]
omnimirror-apt = "omnimirror.cli_apt:main"

